Move components to atoms
